Revert "Send md5 metadata header (#54)"
This reverts commit 2de08beff4feed1367ef9c21ca39497cc0dc5e29.

With the roll-back of the cloud-library from using the AWS S3 lib to
the Minio S3 lib, multi-part uploads to an S3 backend fail with a
403. Minio uploads do succeed.

Local testing shows that reverting sending the md5 hash header gives
working upload against a local library with both S3 and Minio backends.

If the server-side S3 lib is changed from Minio->AWS in future this may
need to be re-added.