refactor: remove tesetify usage
The github.com/stretchr/testify/assert module currently pulls in a
version of the gopkg.in/yaml.v3 module that flags up a vulnerability.
This is used for testing only, and only in an isolated spot. Removing
use of this package in tests cuts the dependencies of this module in
half, which makes it easier to maintain as well as easier to adopt.